Move committing related tests to own spec
